Transformer/Smali2Py: Add transformer up to class annotation
This is a rough transformer which basically transform Smali AST to
the equivalent Python AST (based on ast package). Subtrees are converted
to dicts and tokens (terminals that count) are converted to tuples to
make them easier to read.
Now the work is up to class annotation. Next up will be methods.
TODO: A lot of cleanups is needed in the main.lark to make this transformer
more readable.